check permission for category endpoints

+++ src/main/java/ua/com/obox/dbschema/category/CategoryService.java
@@ -7,6 +7,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import ua.com.obox.authserver.user.UserService;
 import ua.com.obox.dbschema.attachment.AttachmentRepository;
 import ua.com.obox.dbschema.menu.Menu;
 import ua.com.obox.dbschema.menu.MenuRepository;
@@ -18,6 +19,7 @@ import ua.com.obox.dbschema.sorting.EntityOrderRepository;
 import ua.com.obox.dbschema.tools.FieldUpdateFunction;
 import ua.com.obox.dbschema.tools.Validator;
 import ua.com.obox.dbschema.tools.attachment.AttachmentTools;
+import ua.com.obox.dbschema.tools.attachment.ReferenceType;
 import ua.com.obox.dbschema.tools.exception.ExceptionTools;
 import ua.com.obox.dbschema.tools.response.BadFieldsResponse;
 import ua.com.obox.dbschema.tools.response.ResponseErrorMap;
@@ -45,6 +47,7 @@ public class CategoryService {
     private final TranslationRepository translationRepository;
     private final AttachmentRepository attachmentRepository;
     private final UpdateServiceHelper serviceHelper;
+    private final UserService userService;
     private final ResourceBundle translationContent = ResourceBundle.getBundle("translation.messages");
     @Value("${application.image-dns}")
     private String attachmentsDns;
@@ -56,6 +59,7 @@ public class CategoryService {
         AtomicReference<Translation> translation = new AtomicReference<>();
 
         categoryRepository.findByCategoryId(categoryId).orElseThrow(() -> ExceptionTools.notFoundException(".categoryNotFound", finalAcceptLanguage, categoryId));
+        userService.checkPermissionForUser(ReferenceType.category, categoryId, finalAcceptLanguage);
 
         List<Dish> dishes = dishRepository.findAllByCategory_CategoryIdOrderByCreatedAtDesc(categoryId);
 
@@ -106,6 +110,8 @@ public class CategoryService {
         String finalAcceptLanguage = CheckHeader.checkHeaderLanguage(acceptLanguage);
 
         Category category = categoryRepository.findByCategoryId(categoryId).orElseThrow(() -> ExceptionTools.notFoundException(".categoryNotFound", finalAcceptLanguage, categoryId));
+        userService.checkPermissionForUser(ReferenceType.category, categoryId, finalAcceptLanguage);
+
         Translation translation = translationRepository.findAllByTranslationId(category.getTranslationId())
                 .orElseThrow(() -> ExceptionTools.notFoundException(".translationNotFound", finalAcceptLanguage, categoryId));
 
@@ -125,6 +131,7 @@ public class CategoryService {
 
     public CategoryResponseId createCategory(Category request, String acceptLanguage) throws JsonProcessingException {
         String finalAcceptLanguage = CheckHeader.checkHeaderLanguage(acceptLanguage);
+        userService.checkPermissionForUser(ReferenceType.menu, request.getMenuId(), finalAcceptLanguage);
         Map<String, String> fieldErrors = new ResponseErrorMap<>();
 
         Optional<Menu> menu = menuRepository.findByMenuId(request.getMenuId());
@@ -159,9 +166,12 @@ public class CategoryService {
 
     public void patchCategoryById(String categoryId, Category request, String acceptLanguage) throws JsonProcessingException {
         String finalAcceptLanguage = CheckHeader.checkHeaderLanguage(acceptLanguage);
+        userService.checkPermissionForUser(ReferenceType.category, categoryId, finalAcceptLanguage);
         Map<String, String> fieldErrors = new ResponseErrorMap<>();
 
         Category category = categoryRepository.findByCategoryId(categoryId).orElseThrow(() -> ExceptionTools.notFoundException(".categoryNotFound", finalAcceptLanguage, categoryId));
+        userService.checkPermissionForUser(ReferenceType.category, categoryId, finalAcceptLanguage);
+
         Translation translation = translationRepository.findAllByTranslationId(category.getTranslationId())
                 .orElseThrow(() -> ExceptionTools.notFoundException(".translationNotFound", finalAcceptLanguage, categoryId));
 
@@ -175,6 +185,7 @@ public class CategoryService {
     public void deleteCategoryById(String categoryId, String acceptLanguage) {
         String finalAcceptLanguage = CheckHeader.checkHeaderLanguage(acceptLanguage);
         Category category = categoryRepository.findByCategoryId(categoryId).orElseThrow(() -> ExceptionTools.notFoundException(".categoryNotFound", finalAcceptLanguage, categoryId));
+        userService.checkPermissionForUser(ReferenceType.category, categoryId, finalAcceptLanguage);
         categoryRepository.delete(category);
     }
